Rename examples transition components to kebab-case
