+++ packages/examples/src/components/transitions/fade-through/index.module.css.d.ts
