+++ packages/examples/src/components/transitions/fade/index.module.css.d.ts
