+++ packages/examples/src/components/transitions/shared-axis/index.module.css
