+++ packages/examples/src/components/transitions/slide/index.module.css
