+++ packages/examples/src/components/transitions/slide/index.tsx
